requirements.txt
facebook-sdk==3.1.0
tweepy==3.7.0
praw==6.2.0
